backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
python-jose[cryptography]
passlib
email-validator>=1.3.0